Isolate functions for several commands
These changes move the main functionality from the cmd package
(using cobra.Command and args) to the ops package (using a structure)

This change makes it easier to test, and to use dbdeployer as a library from other
packages.
