+++ cmd/delete_binaries.go
@@ -15,69 +15,13 @@
 package cmd
 
 import (
-	"bufio"
 	"fmt"
-	"os"
-	"path"
-	"strings"
 
 	"github.com/datacharmer/dbdeployer/common"
-	"github.com/datacharmer/dbdeployer/defaults"
 	"github.com/datacharmer/dbdeployer/globals"
+	"github.com/datacharmer/dbdeployer/ops"
 	"github.com/spf13/cobra"
 )
-
-func sandboxesUsingBinariesDir(basedir, binariesDir string) []string {
-	var sandboxes []string
-	var sandboxList defaults.SandboxCatalog
-	var err error
-	sandboxList, err = defaults.ReadCatalog()
-	common.ErrCheckExitf(err, 1, "error getting sandboxes from catalog: %s", err)
-	fullPath := path.Join(basedir, binariesDir)
-	for _, sb := range sandboxList {
-		if sb.Origin == fullPath {
-			sandboxes = append(sandboxes, sb.Destination)
-		}
-	}
-	return sandboxes
-}
-
-func deleteBinaries(basedir, binariesDir string, confirm bool) (deleted bool, err error) {
-	fullPath := path.Join(basedir, binariesDir)
-	if !common.DirExists(fullPath) {
-		common.Exitf(1, globals.ErrDirectoryNotFound, fullPath)
-	}
-
-	sandboxesUsingBinaries := sandboxesUsingBinariesDir(basedir, binariesDir)
-	if len(sandboxesUsingBinaries) > 0 {
-		return false, fmt.Errorf("\nbinaries directory %s is used by the following deployments:\n%s",
-			fullPath, strings.Join(sandboxesUsingBinaries, "\n"))
-	}
-	if confirm {
-		common.CondPrintf("Do you want to delete %s? y/[N] ", binariesDir)
-		bio := bufio.NewReader(os.Stdin)
-		line, _, err := bio.ReadLine()
-		if err != nil {
-			fmt.Println(err)
-		} else {
-			answer := string(line)
-			if answer == "y" || answer == "Y" {
-				fmt.Println("Proceeding with deletion")
-			} else {
-				fmt.Println("Deletion skipped at user request")
-				return false, nil
-			}
-		}
-	}
-	_, err = common.RunCmdWithArgs("rm", []string{"-rf", fullPath})
-	if err != nil {
-		return false, fmt.Errorf("error removing %s", fullPath)
-	}
-	if common.DirExists(fullPath) {
-		return false, fmt.Errorf("directory %s was not removed - Reason unknown", fullPath)
-	}
-	return true, nil
-}
 
 func runDeleteBinaries(cmd *cobra.Command, args []string) {
 	if len(args) < 1 {
@@ -92,7 +36,7 @@ func runDeleteBinaries(cmd *cobra.Command, args []string) {
 	basedir, err := getAbsolutePathFromFlag(cmd, "sandbox-binary")
 	common.ErrCheckExitf(err, 1, "error finding absolute path for 'sandbox-binary'")
 
-	isDeleted, err := deleteBinaries(basedir, binariesDir, !skipConfirm)
+	isDeleted, err := ops.DeleteBinaries(basedir, binariesDir, !skipConfirm)
 	if !isDeleted {
 		common.ErrCheckExitf(err, 1, "%s", err)
 		return
