+++ cmd/unpack.go
@@ -17,16 +17,12 @@ package cmd
 
 import (
 	"fmt"
-	"os"
-	"path"
-	"regexp"
-	"strings"
 
+	"github.com/datacharmer/dbdeployer/ops"
 	"github.com/spf13/cobra"
 
 	"github.com/datacharmer/dbdeployer/common"
 	"github.com/datacharmer/dbdeployer/globals"
-	"github.com/datacharmer/dbdeployer/unpack"
 )
 
 func unpackTarball(cmd *cobra.Command, args []string) {
@@ -34,137 +30,35 @@ func unpackTarball(cmd *cobra.Command, args []string) {
 	Basedir, err := getAbsolutePathFromFlag(cmd, "sandbox-binary")
 	common.ErrCheckExitf(err, 1, "error getting absolute path for 'sandbox-binary'")
 	verbosity, _ := flags.GetInt(globals.VerbosityLabel)
+	Prefix, _ := flags.GetString(globals.PrefixLabel)
+	isShell, _ := flags.GetBool(globals.ShellLabel)
+	target, _ := flags.GetString(globals.TargetServerLabel)
+	overwrite, _ := flags.GetBool(globals.OverwriteLabel)
+	flavor, _ := flags.GetString(globals.FlavorLabel)
+	dryRun, _ := flags.GetBool(globals.DryRunLabel)
+	Version, _ := flags.GetString(globals.UnpackVersionLabel)
 	if !common.DirExists(Basedir) {
 		common.Exit(1,
 			fmt.Sprintf(globals.ErrDirectoryNotFound, Basedir),
 			"You should create it or provide an alternate base directory using --sandbox-binary")
 	}
-	tarball := args[0]
-	reVersion := regexp.MustCompile(`(\d+\.\d+\.\d+)`)
-	verList := reVersion.FindAllStringSubmatch(tarball, -1)
 
-	detectedVersion := ""
-	if verList != nil {
-		detectedVersion = verList[0][0]
-	}
-	// common.CondPrintf(">> %#v %s\n",verList, detected_version)
+	err = ops.UnpackTarball(ops.UnpackOptions{
+		SandboxBinary: Basedir,
+		TarballName:   args[0],
+		TargetServer:  target,
+		Version:       Version,
+		Prefix:        Prefix,
+		Flavor:        flavor,
+		Verbosity:     verbosity,
+		IsShell:       isShell,
+		Overwrite:     overwrite,
+		DryRun:        dryRun,
+	})
 
-	isShell, _ := flags.GetBool(globals.ShellLabel)
-	target, _ := flags.GetString(globals.TargetServerLabel)
-	if !isShell && target != "" {
-		common.Exit(1,
-			"unpack: Option --target-server can only be used with --shell")
-	}
-
-	overwrite, _ := flags.GetBool(globals.OverwriteLabel)
-	flavor, _ := flags.GetString(globals.FlavorLabel)
-	dryRun, _ := flags.GetBool(globals.DryRunLabel)
-	if flavor == "" {
-		baseName := common.BaseName(tarball)
-		flavor = common.DetectTarballFlavor(baseName)
-		if flavor == "" {
-			common.Exitf(1, "No flavor detected in %s. Please use --%s", tarball, globals.FlavorLabel)
-		}
-	}
-	Version, _ := flags.GetString(globals.UnpackVersionLabel)
-	if Version == "" {
-		Version = detectedVersion
-	}
-	if Version == "" {
-		common.Exit(1,
-			"unpack: No version was detected from tarball name. ",
-			"Flag --unpack-version becomes mandatory")
-	}
-	// This call used to ensure that the port provided is in the right format
-	_, err = common.VersionToPort(Version)
 	if err != nil {
-		common.Exitf(1, "version %s not in the required format", Version)
+		common.Exitf(1, "error unpacking tarball %s: %s", args[0], err)
 	}
-	Prefix, _ := flags.GetString(globals.PrefixLabel)
-	if isShell {
-		fmt.Printf("%s\n", Version)
-		var canBeEmbedded bool
-		canBeEmbedded, err = common.HasCapability(common.MySQLShellFlavor, common.EmbedMySQLShell, Version)
-		if err != nil {
-			common.Exitf(1, "error detecting shell capability: %s", err)
-		}
-		if !canBeEmbedded {
-			common.Exitf(1, "MySQL shell version %s insufficient for embedding", Version)
-		}
-	}
-
-	destination := path.Join(Basedir, Prefix+Version)
-	if target != "" {
-		destination = path.Join(Basedir, target)
-	}
-	if common.DirExists(destination) && !isShell {
-		if overwrite {
-			if dryRun {
-				fmt.Printf("delete binaries %s %s\n", Basedir, Prefix+Version)
-			} else {
-				isDeleted, err := deleteBinaries(Basedir, Prefix+Version, false)
-				if !isDeleted {
-					common.Exitf(1, "directory %s could not be removed", Prefix+Version)
-				}
-				if err != nil {
-					common.Exitf(1, "error removing directory %s: %s", Prefix+Version, err)
-				}
-			}
-		} else {
-			common.Exitf(1, globals.ErrNamedDirectoryAlreadyExists, "destination directory", destination)
-		}
-	}
-	extracted := path.Base(tarball)
-	var bareName string
-
-	var extractFunc func(string, string, int) error
-	var foundExtension string
-
-	switch {
-	case strings.HasSuffix(tarball, globals.TarGzExt):
-		extractFunc = unpack.UnpackTar
-		foundExtension = globals.TarGzExt
-	case strings.HasSuffix(tarball, globals.TarXzExt):
-		extractFunc = unpack.UnpackXzTar
-		foundExtension = globals.TarXzExt
-	default:
-		common.Exitf(1, "tarball extension must be either '%s' or '%s'", globals.TarGzExt, globals.TarXzExt)
-	}
-	err = unpack.VerifyTarFile(tarball)
-	if err != nil {
-		common.Exitf(1, "validation for %s failed: %s", tarball, err)
-	}
-	bareName = extracted[0 : len(extracted)-len(globals.TarGzExt)]
-	if isShell {
-		common.CondPrintf("Merging shell tarball %s to %s\n", common.ReplaceLiteralHome(tarball), common.ReplaceLiteralHome(destination))
-		if !dryRun {
-			err := unpack.MergeShell(tarball, foundExtension, Basedir, destination, bareName, verbosity)
-			common.ErrCheckExitf(err, 1, "error while unpacking mysql shell tarball : %s", err)
-		}
-		return
-	}
-
-	common.CondPrintf("Unpacking tarball %s to %s\n", tarball, common.ReplaceLiteralHome(destination))
-	if dryRun {
-		return
-	}
-	//verbosity_level := unpack.VERBOSE
-	// err := unpack.UnpackTar(tarball, Basedir, verbosity)
-	err = extractFunc(tarball, Basedir, verbosity)
-	common.ErrCheckExitf(err, 1, "%s", err)
-	finalName := path.Join(Basedir, bareName)
-	// If the directory was not created, it probably means that the tarball was not well organised
-	// and either lacked the top directory or the top directory had a different name
-	if !common.DirExists(finalName) {
-		common.Exitf(1, "problem with tarball %s: directory %s was not created", tarball, finalName)
-	}
-	if finalName != destination {
-		common.CondPrintf("Renaming directory %s to %s\n", finalName, destination)
-		err = os.Rename(finalName, destination)
-		common.ErrCheckExitf(err, 1, "%s", err)
-	}
-	err = common.WriteString(flavor, path.Join(destination, globals.FlavorFileName))
-	common.ErrCheckExitf(err, 1, "error writing %s in %s", globals.FlavorFileName, destination)
 }
 
 // unpackCmd represents the unpack command
