+++ cmd/versions.go
@@ -16,45 +16,12 @@
 package cmd
 
 import (
-	"fmt"
 	"github.com/datacharmer/dbdeployer/globals"
+	"github.com/datacharmer/dbdeployer/ops"
 
 	"github.com/datacharmer/dbdeployer/common"
 	"github.com/spf13/cobra"
 )
-
-func listVersions(dirs []string, basedir, flavor string, iteration int) {
-	maxWidth := 80
-	maxLen := 0
-	for _, dir := range dirs {
-		if len(dir) > maxLen {
-			maxLen = len(dir)
-		}
-	}
-	var header string
-
-	if basedir != "" && iteration == 0 {
-		header = fmt.Sprintf("Basedir: %s\n", basedir)
-	}
-	if flavor != "" {
-		header += fmt.Sprintf("(Flavor: %s)\n", flavor)
-	}
-	if header != "" {
-		fmt.Printf("%s", header)
-	}
-	columns := int(maxWidth / (maxLen + 2))
-	mask := fmt.Sprintf("%%-%ds", maxLen+2)
-	count := 0
-	for _, dir := range dirs {
-		fmt.Printf(mask, dir)
-		count += 1
-		if count > columns {
-			count = 0
-			fmt.Println("")
-		}
-	}
-	fmt.Println("")
-}
 
 // Shows the MySQL versions available in $SANDBOX_BINARY
 // (default $HOME/opt/mysql)
@@ -65,28 +32,13 @@ func showVersions(cmd *cobra.Command, args []string) {
 	flavor, _ := cmd.Flags().GetString(globals.FlavorLabel)
 	byFlavor, _ := cmd.Flags().GetBool(globals.ByFlavorLabel)
 
-	var versionInfoList []common.VersionInfo
-	var dirs []string
-	var flavoredLists = make(map[string][]string)
-
-	versionInfoList = common.GetVersionInfoFromDir(basedir)
-	if byFlavor {
-		for _, verInfo := range versionInfoList {
-			flavoredLists[verInfo.Flavor] = append(flavoredLists[verInfo.Flavor], verInfo.Version)
-		}
-		count := 0
-		for f, versions := range flavoredLists {
-			listVersions(versions, basedir, f, count)
-			count++
-			fmt.Println("")
-		}
-	} else {
-		for _, verInfo := range versionInfoList {
-			if flavor == verInfo.Flavor || flavor == "" {
-				dirs = append(dirs, verInfo.Version)
-			}
-		}
-		listVersions(dirs, basedir, flavor, 0)
+	err = ops.ShowVersions(ops.VersionOptions{
+		SandboxBinary: basedir,
+		Flavor:        flavor,
+		ByFlavor:      byFlavor,
+	})
+	if err != nil {
+		common.Exitf(1, "error showing versions: %s", err)
 	}
 }
