+++ common/COMPATIBLE_VERSION
@@ -1,1 +1,1 @@
-1.62.0
+1.66.0
