+++ common/VERSION
@@ -1,1 +1,1 @@
-1.65.2
+1.66.0
