+++ go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/alexeyco/simpletable v0.0.0-20180729223640-1fa9009f1080
 	github.com/araddon/dateparse v0.0.0-20190622164848-0fb0a474d195
 	github.com/dustin/go-humanize v1.0.0
+	github.com/frankban/quicktest v1.14.3
 	github.com/go-sql-driver/mysql v1.4.1
 	github.com/nightlyone/lockfile v0.0.0-20180618180623-0ad87eef1443
 	github.com/olekukonko/tablewriter v0.0.5 // indirect
