+++ unpack/unpack-shell.go
@@ -41,7 +41,7 @@ func MergeShell(tarball, extension, basedir, destination, bareName string, verbo
 	for _, dir := range dirs {
 		destPath := path.Join(destination, dir)
 		if !common.DirExists(destPath) {
-			common.Exitf(1, "destination server directory %s does not exist in %s\n", dir, destination)
+			return fmt.Errorf("destination server directory %s does not exist in %s", dir, destination)
 		}
 		destPath = path.Join(destination, dir, "mysqlsh")
 		if dir != "bin" && common.DirExists(destPath) {
